requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
google-generativeai==0.7.2
openai==1.40.0
pdfplumber==0.11.3
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==10.4.0
pandas==2.2.2
python-dotenv==1.0.1
